gateway/requirements.txt
pyserial>=3.5
pycryptodome>=3.18
pytest>=7.0